Fixing api call and cementing my framework
Abandoned MvvmCross Idea. It doesn't make sense to use it in Xamarin.Forms

Fixed issue in view model - VM was making a weird call to an async method and it wasn't set to await as it was being called from a non-async context. Using the override Initialize method seems to be much better as it gives me an async context to call the api from.